# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BNB Management API",
        "description": "Professional BNB Management System API with JWT Authentication",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "https://3a3b2bcb1e34.ngrok-free.app/api/v1",
            "description": "Public Tunnel (ngrok)"
        }
    ],
    "paths": {
...
        "/bnbs/{bnb}/availability": {
            "get": {
                "tags": [
                    "Availability"
                ],
                "summary": "Get availability calendar",
                "description": "Get availability calendar for a date range with pricing information.",
                "operationId": "getBNBAvailabilityCalendar",
                "parameters": [
                    {
                        "name": "bnb",
                        "in": "path",
                        "description": "BNB property ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Start date for calendar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-01"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "End date for calendar",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2025-12-31"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Availability calendar retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Availability calendar retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "bnb": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "name": {
                                                            "type": "string",
                                                            "example": "Downtown Apartment"
                                                        },
                                                        "default_price": {
                                                            "type": "string",
                                                            "example": "150.00"
                                                        },
                                                        "default_availability": {
                                                            "type": "boolean",
                                                            "example": true
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "calendar": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2025-12-01"
                                                            },
                                                            "is_available": {
                                                                "type": "boolean",
                                                                "example": true
                                                            },
                                                            "price_override": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": null,
                                                                "nullable": true
                                                            },
                                                            "effective_price": {
                                                                "type": "string",
                                                                "example": "150.00"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "BNB not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "BNB not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bnbs/{bnb}/availability/block": {
            "post": {
                "tags": [
                    "Availability"
                ],
                "summary": "Block dates",
                "description": "Block multiple dates at once making them unavailable for booking.",
                "operationId": "blockBNBDates",
                "parameters": [
                    {
                        "name": "bnb",
                        "in": "path",
                        "description": "BNB property ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "start_date",
                                    "end_date"
                                ],
                                "properties": {
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-24"
                                    },
                                    "end_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-12-26"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Holiday period - not available",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dates blocked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dates blocked successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "blocked_dates": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2025-12-24"
                                                    }
                                                },
                                                "reason": {
                                                    "type": "string",
                                                    "example": "Holiday period - not available"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "BNB not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "BNB not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}